Getting Ready for Comp
Created an auto opmode. This opmode will move the robot forward at 0.3 power for 0.8 seconds, and then stop. This follows the same motor naming conventions as we have elsewhere.

Created a class for field centric driving. This uses the robots internal Inertial Measurement unit to calculate how it needs to drive. This will make the robot much easier to drive than our earlier robot centric driving

Added a program for the spindexer, currently it only sorts for colors, but that could change in the future if we continue with the spindexer.

Modified the flywheel class to change it from an opmode just to a class. Also changed how it works. It is now the turret class, and has been changed to always run the flywheel at 0.5 speed and increase it to 1.0 when the left bumper is pressed. We have also added the intake, which will run at 0.3 speed and increase to 0.6 when the right bumper is pressed. Note this uses gamepad2.

Added a CompTeleop opmode. Currently this uses the MecanumDrive and Turret classes to run the robot. This consolidates the classes into one opmode.

:)